refactor(client): moving the schema to folder
